# Dupe and escape span values only if necessary