@jneen 2019-07-03
Would you mind benchmarking on a very long comment that contains escape characters?