@ashmaroli 2019-07-03
> Would you mind benchmarking on a very long comment

If you provide a sample string, I shall use that.