@jneen 2019-07-03
You can find some fairly long comments in `spec/visual/samples`, maybe add a `&` inside the comment to the end of it? My concern is that for large single tokens we will scan the whole thing twice. It's possible this is very low impact, but I want to make sure.