@ashmaroli 2019-07-03
@jneen Okay, updated the above benchmark script with the following:
```ruby
require 'yaml'

# set up a long string text

PLAINTEXT = YAML.load(<<-TEXT)
comment: >
  @param [Proc] callback a block that will be evaluated in the context of the lexer
  if `re` matches. This block has access to a number of lexer methods, including
  `RegexLexer#push`, `RegexLexer#pop!`, `RegexLexer#token` et `RegexLexer#delegate`.
  The first argument can be used to access the match groups.
  -endcomment-
TEXT

RAWTEXT = YAML.load(<<-TEXT)
comment: >
  @param [Proc] callback a block that will be evaluated in the context of the lexer
  if `re` matches. This block has access to a number of lexer methods, including
  #<RegexLexer#push>, #<RegexLexer#pop!>, #<RegexLexer#token>, & #<RegexLexer#delegate>.
  The first argument can be used to access the match groups.
  <endcomment>
TEXT

[
  ['plain', PLAINTEXT['comment']],
  ['specl', RAWTEXT['comment']]
].each do |key, value|
  Benchmark.ips do |x|
    ...
```
The result are similar to previous ones:
```
Warming up --------------------------------------
           esc plain    39.150k i/100ms
      cond_esc plain    41.220k i/100ms
Calculating -------------------------------------
           esc plain    603.987k (A± 0.7%) i/s -      3.054M in   5.056187s
      cond_esc plain    666.529k (A± 0.5%) i/s -      3.339M in   5.009407s

Comparison:
      cond_esc plain:   666529.2 i/s
           esc plain:   603986.5 i/s - 1.10x  slower

Warming up --------------------------------------
           esc specl     6.219k i/100ms
      cond_esc specl     5.802k i/100ms
Calculating -------------------------------------
           esc specl     68.860k (A± 0.6%) i/s -    348.264k in   5.057804s
      cond_esc specl     64.418k (A± 0.6%) i/s -    324.912k in   5.044026s

Comparison:
           esc specl:    68859.7 i/s
      cond_esc specl:    64417.6 i/s - 1.07x  slower
```