@jneen 2019-07-03
Okay! I was thinking about the occasional pages-long comment that you will see on gitlab, etc, but honestly I think that's rare enough that the tradeoff is worth it. If you could add a comment explaining that it's done that way for performance reasons, then it's +1 to merge from me.